Add files via upload
This code allows to obtain sectors or clusters of boats by using unsupervised machine learning techniques, in this case the K-means classification algorithm is applied. It is an unsupervised classification algorithm (clustering) that groups objects into k groups based on their characteristics. The clustering is performed by minimizing the sum of distances between each object and the centroid of its group or cluster. The quadratic distance is usually used.

The algorithm consists of three steps:

Initialization: once the number of clusters, k, is chosen using the Elbow method, k centroids are established in the data space, e.g., by choosing them randomly.
Assigning objects to centroids: each object in the data is assigned to its nearest centroid.
Updating centroids: the position of the centroid of each group is updated by taking as the new centroid the position of the average of the objects belonging to that group.